Исправления
Всё что в ООП было написано - исправил
