+++ oop/Library/Book.java
@@ -1,7 +1,10 @@
+package Library;
+
+
 public class Book {
-    String title;
-    Author author;
-    int year;
+    private String title;
+    private Author author;
+    private int year;
 
     public Book(String title, Author author, int year) {
         this.setTitle(title);
@@ -38,17 +41,4 @@ public class Book {
         return "Title: %s, Author: %s, Year: %d\n".formatted(this.title, this.author.toString(), this.year);
     }
 
-    static class Main {
-        public static void main(String[] args) throws Exception {
-            Author author = new Author("Alexander Pushkin", "male", "pushkin.alex@duelloser.ru");
-            Book book = new Book("Captain's daughter", author, 1836);
-
-            System.out.printf("%s, %s, %s\n\n", author.getName(), author.getGender(), author.getEmail());
-            System.out.printf("%s, %s, %s\n\n", book.getTitle(), book.getAuthor().toString(), book.getYear());
-
-            System.out.println(author.toString());
-            System.out.println();
-            System.out.println(book.toString());
-        }
-    }
 }
