+++ oop/Pizza/Size.java
@@ -1,0 +1,5 @@
+package Pizza;
+
+public enum Size {
+    SMALL, MEDIUM, BIG
+}
